Update README.md
Added the following
" ## Creator: Nickolas Duong
## Version: 1.1 (Java)
## Date of Creation: 2/12/20
## Project Description
This project was created to assist ice cream parlors with getting orders from customers online rather than in person. This is to allow the parlors to be found online rather than by a chance visit.
## Project Purpose
The purpose of the project is to make it easy for an ice cream parlor and their inventory to be added to the program for customers to see. It is also designed to allow for multiple parlors to be on the same system instead of each one developing their own personal system. Its other purpose is to allow the customer to find an ice cream parlor near them to place the order at.
## Program Usage
### 1. Open cmd console or equivalent
### 2. Enter directory master directory of the program
### 3. Input "java parlororder/IceCreamForU" to run the program
### 4. Follow the program, only inputting numbers or a single letter when asked "